docs: updated documentation
