+++ .containerignore
@@ -2,7 +2,7 @@
 data/
 chroma/
 models/
-Python-3.12.9.tgz  # Archivo innecesario si usas la imagen base de Python
+Python-3.12.9.tgz
 
 # Ignorar archivos de entorno locales
 .env
