+++ _config.yml
@@ -93,14 +93,12 @@ exclude:
   - logs/
   - tests/
   - unitTests/
-  - prometheus/
   - .env
   - .env.example
   - "*.py"
   - "*.sh"
   - "*.yml"
   - "*.yaml"
-  - Containerfile
   - Dockerfile
   - requirements.txt
   - pytest.ini
