+++ app/requirements.txt
@@ -1,0 +1,220 @@
+accelerate==1.9.0
+aiosqlite==0.21.0
+annotated-types==0.7.0
+anyio==4.9.0
+asttokens==3.0.0
+attrs==25.3.0
+backcall==0.2.0
+backoff==2.2.1
+bcrypt==4.3.0
+beautifulsoup4==4.13.4
+build==1.2.2.post1
+cachetools==5.5.2
+calmap==0.0.11
+certifi==2025.7.14
+charset-normalizer==3.4.2
+click==8.2.1
+coloredlogs==15.0.1
+contourpy==1.3.2
+cycler==0.12.1
+decorator==5.2.1
+dill==0.4.0
+distro==1.9.0
+docker==7.1.0
+durationpy==0.10
+easyocr==1.7.2
+et-xmlfile==2.0.0
+executing==2.2.0
+fastapi==0.116.1
+filelock==3.18.0
+filetype==1.2.0
+flatbuffers==25.2.10
+fonttools==4.59.0
+fsspec==2025.7.0
+google-ai-generativelanguage==0.6.18
+google-api-core==2.25.1
+google-auth==2.40.3
+googleapis-common-protos==1.70.0
+greenlet==3.2.3
+grpcio==1.73.1
+grpcio-status==1.73.1
+h11==0.16.0
+hf-xet==1.1.5
+httpcore==1.0.9
+httptools==0.6.4
+httpx==0.28.1
+huggingface-hub==0.33.4
+humanfriendly==10.0
+idna==3.10
+imageio==2.37.0
+importlib-metadata==8.7.0
+importlib-resources==6.5.2
+iniconfig==2.1.0
+ipython==8.12.3
+jedi==0.19.2
+jinja2==3.1.6
+jiter==0.10.0
+jsonlines==3.1.0
+jsonpatch==1.33
+jsonpointer==3.0.0
+jsonref==1.1.0
+jsonschema==4.25.0
+jsonschema-specifications==2025.4.1
+kiwisolver==1.4.8
+kubernetes==33.1.0
+langchain==0.3.26
+langchain-core==0.3.71
+langchain-google-genai==2.1.8
+langchain-openai==0.3.28
+langchain-text-splitters==0.3.8
+langgraph==0.5.4
+langgraph-checkpoint==2.1.1
+langgraph-checkpoint-sqlite==2.0.11
+langgraph-prebuilt==0.5.2
+langgraph-sdk==0.1.74
+langsmith==0.4.8
+langfuse==3.4.0
+latex2mathml==3.78.0
+lazy-loader==0.4
+lxml==5.4.0
+markdown-it-py==3.0.0
+marko==2.1.4
+markupsafe==3.0.2
+matplotlib==3.10.3
+matplotlib-inline==0.1.7
+mdurl==0.1.2
+mmh3==5.1.0
+mpire==2.10.2
+mpmath==1.3.0
+multiprocess==0.70.18
+networkx==3.4.2
+ninja==1.11.1.4
+numpy==2.2.6
+nvidia-cublas-cu12==12.6.4.1
+nvidia-cuda-cupti-cu12==12.6.80
+nvidia-cuda-nvrtc-cu12==12.6.77
+nvidia-cuda-runtime-cu12==12.6.77
+nvidia-cudnn-cu12==9.5.1.17
+nvidia-cufft-cu12==11.3.0.4
+nvidia-cufile-cu12==1.11.1.6
+nvidia-curand-cu12==10.3.7.77
+nvidia-cusolver-cu12==11.7.1.2
+nvidia-cusparse-cu12==12.5.4.2
+nvidia-cusparselt-cu12==0.6.3
+nvidia-nccl-cu12==2.26.2
+nvidia-nvjitlink-cu12==12.6.85
+nvidia-nvtx-cu12==12.6.77
+oauthlib==3.3.1
+onnxruntime==1.22.1
+openai==1.97.1
+opencv-python-headless==4.11.0.86
+openpyxl==3.1.5
+opentelemetry-api==1.35.0
+opentelemetry-exporter-otlp-proto-common==1.35.0
+opentelemetry-exporter-otlp-proto-grpc==1.35.0
+opentelemetry-proto==1.35.0
+opentelemetry-sdk==1.35.0
+opentelemetry-semantic-conventions==0.56b0
+orjson==3.11.0
+ormsgpack==1.10.0
+overrides==7.7.0
+packaging==25.0
+pandas==2.3.1
+parso==0.8.4
+pexpect==4.9.0
+pickleshare==0.7.5
+pillow==11.3.0
+pluggy==1.6.0
+posthog==5.4.0
+prompt-toolkit==3.0.51
+proto-plus==1.26.1
+protobuf==6.31.1
+psutil==7.0.0
+ptyprocess==0.7.0
+pure-eval==0.2.3
+pyasn1==0.6.1
+pyasn1-modules==0.4.2
+pybase64==1.4.1
+pyclipper==1.3.0.post6
+pydantic==2.11.7
+pydantic-core==2.33.2
+pydantic-settings==2.10.1
+pydantic[email]
+pygments==2.19.2
+pylatexenc==2.10
+pyparsing==3.2.3
+pypdfium2==4.30.0
+pypika==0.48.9
+pyproject-hooks==1.2.0
+pytest==8.4.1
+pytest-asyncio==1.1.0
+pytest-dotenv==0.5.2
+pytest-testinfra==10.2.2
+python-bidi==0.6.6
+python-dateutil==2.9.0.post0
+python-docx==1.2.0
+python-dotenv==1.1.1
+python-multipart==0.0.20
+python-pptx==1.0.2
+pytz==2025.2
+pyyaml==6.0.2
+referencing==0.36.2
+regex==2024.11.6
+requests==2.32.4
+requests-oauthlib==2.0.0
+requests-toolbelt==1.0.0
+rich==14.0.0
+rpds-py==0.26.0
+rsa==4.9.1
+rtree==1.4.0
+safetensors==0.5.3
+scikit-image==0.25.2
+scipy==1.15.3
+seaborn==0.13.2
+semchunk==2.2.2
+setuptools==80.9.0
+shapely==2.1.1
+shellingham==1.5.4
+six==1.17.0
+sniffio==1.3.1
+soupsieve==2.7
+sqlalchemy==2.0.41
+sqlite-vec==0.1.6
+stack-data==0.6.3
+starlette==0.47.2
+sympy==1.14.0
+tabulate==0.9.0
+tenacity==9.1.2
+testinfra==6.0.0
+tifffile==2025.5.10
+tiktoken==0.9.0
+tokenizers==0.21.2
+torch==2.7.1
+torchvision==0.22.1
+tqdm==4.67.1
+traitlets==5.14.3
+transformers==4.53.3
+triton==3.3.1
+typer==0.16.0
+typing-extensions==4.14.1
+typing-inspection==0.4.1
+tzdata==2025.2
+urllib3==2.5.0
+uvicorn==0.35.0
+uvloop==0.21.0
+watchfiles==1.1.0
+wcwidth==0.2.13
+websocket-client==1.8.0
+websockets==15.0.1
+wikipedia==1.4.0
+xlsxwriter==3.2.5
+xxhash==3.5.0
+zipp==3.23.0
+zstandard==0.23.0
+
+# LTI 1.3 Integration
+PyLTI1p3==2.0.0
+PyJWT==2.10.1
+cryptography==46.0.2
+jwcrypto==1.5.6
+motor==3.6.0
