--- current-url.txt
@@ -1,1 +1,0 @@
-https://prophet-defined-sample-monitors.trycloudflare.com
