--- prometheus/docker-compose-prometheus-graphana.yml
@@ -1,19 +1,0 @@
-# docker-compose.yaml
-version: "3"
-
-services:
-  prometheus:
-    image: docker.io/prom/prometheus:latest
-    extra_hosts:
-      - "host.docker.internal:host-gateway"     # allow a direct connection from container to the local machine
-    ports:
-      - "9091:9090"   # the default port used by Prometheus
-    volumes:
-      - ./prometheus.yml:/etc/prometheus/prometheus.yml:z # mount Prometheus config file
-
-  grafana:
-    image: docker.io/grafana/grafana:latest
-    depends_on:
-      - prometheus
-    ports:
-      - "3001:3000" # the default port used by Grafana
