--- prometheus/prometheus.yml
@@ -1,10 +1,0 @@
-# prometheus.yaml
-global:
-  scrape_interval: 5s
-  evaluation_interval: 30s
-
-scrape_configs:
-  - job_name: vllm
-    static_configs:
-      - targets:
-          - 'host.docker.internal:8001'
