--- unitTests/__init__.py
@@ -1,1 +1,0 @@
-# Unit tests package
