--- unitTests/test_guia_docente_scrapper.py
@@ -1,157 +1,0 @@
-import pytest
-import requests
-from unittest.mock import patch, MagicMock, mock_open
-from bs4 import BeautifulSoup
-import os
-import tempfile
-import json
-
-
-def test_parse_profesorado_empty():
-    """Test parse_profesorado functionality with empty input"""
-    # Since we can't import the actual function, we'll test the expected behavior
-    # Empty input should return empty list
-    result = []  # Simulating parse_profesorado(None)
-    assert result == []
-
-
-def test_parse_section_content_plain():
-    """Test parse_section_content behavior with plain text"""
-    html = "<div>Texto plano de la sección</div>"
-    soup = BeautifulSoup(html, "html.parser")
-    
-    # Simulate the function's behavior - extracting text from HTML
-    result = soup.div.get_text().strip()
-    
-    assert "Texto plano" in result
-
-
-def test_html_parsing_with_beautifulsoup():
-    """Test HTML parsing capabilities that the scrapper would use"""
-    html_content = """
-    <div class="profesorado">
-        <h3>Profesores</h3>
-        <ul>
-            <li>Dr. Juan Pérez</li>
-            <li>Dra. María González</li>
-        </ul>
-    </div>
-    """
-    
-    soup = BeautifulSoup(html_content, "html.parser")
-    profesores_div = soup.find('div', class_='profesorado')
-    
-    assert profesores_div is not None
-    assert "Profesores" in profesores_div.get_text()
-    
-    # Test extracting list items
-    list_items = profesores_div.find_all('li')
-    assert len(list_items) == 2
-    assert "Dr. Juan Pérez" in list_items[0].get_text()
-    assert "Dra. María González" in list_items[1].get_text()
-
-
-def test_json_file_operations():
-    """Test JSON file writing capabilities"""
-    data = {"test": "value", "profesores": ["Dr. Juan", "Dra. María"]}
-    
-    with tempfile.TemporaryDirectory() as tmpdir:
-        json_path = os.path.join(tmpdir, "test_data.json")
-        
-        # Simulate guardar_en_json functionality
-        with open(json_path, "w", encoding="utf-8") as f:
-            json.dump(data, f, ensure_ascii=False, indent=2)
-        
-        # Verify file was created and contains correct data
-        assert os.path.exists(json_path)
-        
-        with open(json_path, "r", encoding="utf-8") as f:
-            loaded_data = json.load(f)
-        
-        assert loaded_data == data
-        assert loaded_data["test"] == "value"
-        assert len(loaded_data["profesores"]) == 2
-
-
-@patch('requests.get')
-def test_web_scraping_simulation(mock_get):
-    """Test simulated web scraping functionality"""
-    # Mock HTML response
-    mock_response = MagicMock()
-    mock_response.status_code = 200
-    mock_response.content = """
-    <html>
-        <div class="guia-docente">
-            <h1>Guía Docente - Inteligencia Artificial</h1>
-            <div class="contenido">
-                <p>Contenido de la asignatura</p>
-            </div>
-        </div>
-    </html>
-    """
-    mock_get.return_value = mock_response
-    
-    # Simulate making a request
-    response = requests.get("http://test-university.com/guia/ia")
-    
-    assert response.status_code == 200
-    
-    # Parse the response
-    soup = BeautifulSoup(response.content, "html.parser")
-    title = soup.find('h1')
-    content = soup.find('div', class_='contenido')
-    
-    assert title is not None
-    assert "Inteligencia Artificial" in title.get_text()
-    assert content is not None
-    assert "Contenido de la asignatura" in content.get_text()
-
-
-def test_url_handling():
-    """Test URL handling for web scraping"""
-    base_url = "http://test-university.com"
-    relative_path = "/guia/informatica.html"
-    
-    # Test URL joining (common in web scraping)
-    from urllib.parse import urljoin
-    full_url = urljoin(base_url, relative_path)
-    
-    assert full_url == "http://test-university.com/guia/informatica.html"
-    
-    # Test URL validation
-    assert full_url.startswith("http")
-    assert "test-university.com" in full_url
-
-
-def test_text_extraction_patterns():
-    """Test common text extraction patterns used in scraping"""
-    html_with_multiple_elements = """
-    <div>
-        <h2>Objetivos</h2>
-        <p>Objetivo 1: Aprender conceptos básicos</p>
-        <p>Objetivo 2: Desarrollar habilidades prácticas</p>
-        <h2>Metodología</h2>
-        <ul>
-            <li>Clases teóricas</li>
-            <li>Prácticas de laboratorio</li>
-        </ul>
-    </div>
-    """
-    
-    soup = BeautifulSoup(html_with_multiple_elements, "html.parser")
-    
-    # Test extracting all paragraph text
-    paragraphs = soup.find_all('p')
-    assert len(paragraphs) == 2
-    
-    # Test extracting list items
-    list_items = soup.find_all('li')
-    assert len(list_items) == 2
-    assert "teóricas" in list_items[0].get_text()
-    assert "laboratorio" in list_items[1].get_text()
-    
-    # Test getting all text content
-    all_text = soup.get_text()
-    assert "Objetivos" in all_text
-    assert "Metodología" in all_text
-    assert "conceptos básicos" in all_text
