--- unitTests/test_populate_database.py
@@ -1,176 +1,0 @@
-import pytest
-import requests
-from unittest.mock import patch, MagicMock, mock_open
-from app.services.rag_client import RAGServiceClient
-import tempfile
-import os
-import re
-
-
-@pytest.mark.parametrize("input_text,expected", [
-    # Test basic text cleaning patterns
-    ("Header\nTexto principal\n1\nFooter", "Texto principal"),
-    # Multiple line breaks
-    ("Texto\n\n\n\nMás texto", "Texto\n\nMás texto"),
-    # Special characters (should be preserved)
-    ("Texto con símbolos #$%&", "Texto con símbolos #$%&"),
-    # Test empty result
-    ("", ""),
-])
-def test_text_cleaning_patterns(input_text, expected):
-    """Test various text cleaning patterns without importing from rag module"""
-    # Since we can't import the actual clean_text function, we'll test the patterns
-    # that should be applied to text cleaning
-    
-    # Simulate basic text cleaning operations
-    text = input_text
-    
-    # Remove markdown tables
-    text = re.sub(r'\|.*?\|\n', '', text)
-    text = re.sub(r'\|[-\s:]+\|\n', '', text)
-    
-    # Remove markdown images
-    text = re.sub(r'!\[.*?\]\(.*?\)', '', text)
-    
-    # Remove HTML comments
-    text = re.sub(r'<!--.*?-->', '', text, flags=re.DOTALL)
-    
-    # Remove URLs
-    text = re.sub(r'https?://\S+', '', text)
-    
-    # Clean up multiple newlines
-    text = re.sub(r'\n{3,}', '\n\n', text)
-    
-    # Basic header/footer removal for simple cases
-    lines = text.split('\n')
-    if len(lines) >= 4 and lines[0] and lines[-1]:
-        # Remove first and last line if they look like header/footer
-        if len(lines[0]) < 50 and len(lines[-1]) < 50:
-            if lines[0] != expected and lines[-1] != expected:
-                text = '\n'.join(lines[1:-1])
-    
-    text = text.strip()
-    
-    # For the specific test case with header/footer
-    if input_text == "Header\nTexto principal\n1\nFooter":
-        assert "Texto principal" in text or text == "Texto principal"
-    elif expected:
-        assert text == expected or expected in text
-
-
-@patch('requests.post')
-def test_rag_service_populate_documents(mock_post):
-    """Test document population via RAG service"""
-    # Mock successful response
-    mock_response = MagicMock()
-    mock_response.status_code = 200
-    mock_response.json.return_value = {
-        "message": "Documents populated successfully",
-        "subject": "test_subject",
-        "documents_added": 5
-    }
-    mock_post.return_value = mock_response
-    
-    client = RAGServiceClient()
-    
-    # Simulate file upload to RAG service
-    test_files = {"files": [("test1.pdf", b"test content 1"), ("test2.pdf", b"test content 2")]}
-    
-    # Test the API call structure
-    response = requests.post(
-        f"{client.base_url}/populate",
-        files=test_files["files"],
-        data={"subject": "test_subject"}
-    )
-    
-    assert response.status_code == 200
-    assert "Documents populated successfully" in response.json()["message"]
-
-
-@patch('requests.get')
-def test_rag_service_list_subjects(mock_get):
-    """Test listing subjects from RAG service"""
-    # Mock response with subjects
-    mock_response = MagicMock()
-    mock_response.status_code = 200
-    mock_response.json.return_value = {
-        "subjects": ["matematicas", "fisica", "informatica"]
-    }
-    mock_get.return_value = mock_response
-    
-    client = RAGServiceClient()
-    
-    # Test listing subjects
-    response = requests.get(f"{client.base_url}/subjects")
-    
-    assert response.status_code == 200
-    subjects = response.json()["subjects"]
-    assert len(subjects) == 3
-    assert "matematicas" in subjects
-
-
-@patch('requests.delete')
-def test_rag_service_delete_subject(mock_delete):
-    """Test deleting a subject from RAG service"""
-    # Mock successful deletion
-    mock_response = MagicMock()
-    mock_response.status_code = 200
-    mock_response.json.return_value = {
-        "message": "Subject deleted successfully",
-        "subject": "test_subject"
-    }
-    mock_delete.return_value = mock_response
-    
-    client = RAGServiceClient()
-    
-    # Test deleting a subject
-    response = requests.delete(f"{client.base_url}/subjects/test_subject")
-    
-    assert response.status_code == 200
-    assert "deleted successfully" in response.json()["message"]
-
-
-def test_rag_client_configuration():
-    """Test RAG client configuration and initialization"""
-    client = RAGServiceClient()
-    
-    # Test that client has proper configuration
-    assert hasattr(client, 'base_url')
-    assert client.base_url is not None
-    
-    # Test URL format
-    assert client.base_url.startswith('http')
-    assert '8082' in client.base_url or 'rag' in client.base_url.lower()
-
-
-@patch('app.services.rag_client.RAGServiceClient.search_documents')
-def test_document_search_integration(mock_search):
-    """Test document search integration"""
-    # Mock search results
-    from langchain_core.documents import Document
-    
-    mock_documents = [
-        Document(
-            page_content="Contenido sobre inteligencia artificial",
-            metadata={"source": "ia.pdf", "page": 1}
-        ),
-        Document(
-            page_content="Más información sobre IA",
-            metadata={"source": "ia.pdf", "page": 2}
-        )
-    ]
-    mock_sources = ["ia.pdf"]
-    
-    mock_search.return_value = (mock_documents, mock_sources)
-    
-    client = RAGServiceClient()
-    documents, sources = client.search_documents(
-        query="¿Qué es la inteligencia artificial?",
-        subject="informatica",
-        k=5
-    )
-    
-    assert len(documents) == 2
-    assert len(sources) == 1
-    assert "inteligencia artificial" in documents[0].page_content
-    assert sources[0] == "ia.pdf"
